readme move to bregman
